Call celery directly from supervisor.
Having the intermediate shell script causes the signals to not
propagate correctly.  So when supervisor stops the workers the
celery root process dies but the actual workers hang around. On
sandboxes this results in there eventually being lots of stray
workers that eventually chew through all the memory on the box.