The sed -E still doesn't work on OS-X
Noticed when something that worked for Feanil didn't work for Michael

In 7cd04db1 this was updated to rely less on gnu sed, but it still
didn't work properly on the sed shipped with OS-X Sierra.  gsed does,
but I don't want people to need to install that.

Error was that (:\S*)? was only matching : not :tagname, so the
generated Docker file was foo:testbar rather than foo:test